Refactor, reorganize modules
Remove common.py and file_handling.py. New directory common. Move all
classes in deleted modules to the new modules single_file_read_write,
directory_handling, and base.
