+++ tests/integration/chiasma_validation_deploy_tests.py
@@ -2,7 +2,7 @@ import unittest
 from unittest import skip
 from release_ccharp.snpseq_workflow import SnpseqWorkflow
 from release_ccharp.apps.chiasma import Application
-from release_ccharp.apps.common import WindowsCommands
+from release_ccharp.apps.common_xxx import WindowsCommands
 from release_ccharp.utility.os_service import OsService
